// src/app/model/plan.ts
export class Plan {
    id: number = 0;
    cliente: string = '';
    vehiculo: string = '';
    moneda: string = '';
    tipoCambio: number = 0;
    porcentajeCuotaInicial: number = 0;
    cuotaInicial: number = 0;
    montoPrestamo: number = 0;
    porcentajeCuotaFinal: number = 0;
    precioVenta: number = 0;
    costosIniciales: number = 0;
    cuotaFinal: number = 0;
    porcentajeFinancEnCuotas: number = 0;
    financEnCuotas: number = 0;
    tipoTasa: string = '';
    tasa: number = 0;
    plan: number = 0;
    seguroDesgravamen: string = '';
    tipoPeriodoGracia: string = '';
    periodosGracia: number = 0;
    // tasaDescuento: number = 0;
  }
  